+++ game/bishop.py
@@ -1,5 +1,8 @@
 from game.piece import Piece
 
 class Bishop(Piece):
-    def __init__(self, color):
-        super().__init__(color)
+    def __str__(self):
+        if self.color == "WHITE":
+            return "♗"
+        else:
+            return "♝"