+++ game/piece.py
@@ -4,3 +4,4 @@ class Piece:
 
     def movimientos_validos(self, board):
         raise NotImplementedError("Este método debe ser implementado por las subclases")
+