+++ Changelog.md
@@ -1,13 +1,30 @@
 # Changelog
 
+## 30/08/2024
+
+- Move method was not implemented at this time, but it will be added in the future.
+
+- Added the following methods whit their tests in chess:
+    init
+    is_playing
+    turn
+    change_turn
+    show_board
+
+-Exeptions added to board with tests  
+
+## 27/08/2024
+- Test board 100% coveredS
+- First real working commit
+
 ## 18/08/2024
 
--main comented
--Rook comented
--Queen constructor
--King constructor
--Knight constructor
--bishop constructor
+- main comented
+- Rook comented
+- Queen constructor
+- King constructor
+- Knight constructor
+- bishop constructor
 
 ## 17/08/2024
 -Added kings & queen positions to the board and tests for every piece actually exiting in board